requirements.txt
torch>=1.13.0
torchvision>=0.14.0
pandas
numpy
nltk
matplotlib
tqdm
Pillow
wordcloud
scikit-learn
